requirements.txt
# Require python 3.8

# Core
numpy
pyserial
newportESP
pyyaml

# Documentation
sphinx
myst-parser
furo
linkify-it-py
sphinx-copybutton